backend/requirements.txt
blinker==1.6.2
click==8.1.5
colorama==0.4.6
Flask==2.3.2
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
python-dotenv==1.0.0
Werkzeug==2.3.6